Trac #32061: allow gengetopt to build without makeinfo.
The gengetopt release tarball has a bug wherein the maintainer-only
rebuild rules for doc/gengetopt.info and doc/gengetopt.html get
triggered, and fail if the end user does not have the "makeinfo"
program installed. This is not by design; to work around it, we
"touch" the pre-generated files in the middle of the build (if it
fails the first time around...) to avoid rebuilding them.